move GTM, HubSpot, and PostHog from root layout to (site)/layout
