+++ frontends/main/src/app/(site)/SiteProviders.tsx
@@ -1,0 +1,12 @@
+"use client"
+
+import React from "react"
+import ConfiguredPostHogProvider from "@/page-components/ConfiguredPostHogProvider/ConfiguredPostHogProvider"
+
+export default function SiteProviders({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <ConfiguredPostHogProvider>{children}</ConfiguredPostHogProvider>
+}
